requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
anyio==4.6.2.post1
argh==0.26.2
asgiref==3.9.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.0
black==23.12.0
boto3==1.35.15
botocore==1.35.15
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.2.0
channels_redis==4.2.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.5.4
crispy-bootstrap5==2024.2
cryptography==43.0.0
daphne==4.2.1
defusedxml==0.7.1
Django==5.2
django-allauth==0.61.1
django-appconf==1.0.6
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-celery-email==3.0.0
django-cors-headers==4.4.0
django-countries==7.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-drf-filepond==0.5.0
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.1
django-formtools==2.5.1
django-modeltranslation==0.19.17
django-parler==2.3
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-rosetta==0.10.2
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==21.0.0
flake8==3.9.2
flower==2.0.1
frozenlist==1.4.1
googletrans==3.1.0a0
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httptools==0.6.4
httpx==0.13.3
humanize==4.10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.9.3
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.4.0
mccabe==0.6.1
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.43
Pillow==10.0.0
platformdirs==4.2.2
pluggy==1.5.0
polib==1.2.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.7.0
pycparser==2.22
pyflakes==2.3.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==5.3.1
redis==5.0.7
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rpds-py==0.27.1
s3transfer==0.10.2
service-identity==24.2.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tomli==2.0.1
tornado==6.4.1
twilio==9.9.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchdog==1.0.2
watchfiles==1.0.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.9.4
zope.interface==7.2
